Run end-to-end tests on CI
This will run end-to-end tests on the deployed dev environment to ensure
we're testing the entire end-to-end journey with a real front and back
end to catch anything that might break functionality.

The script clones the e2e test repo locally, installs dependencies and
runs the Playwright test script.

Log in secrets are stored in the Circle CI env vars.