Introduce a HttpMessageConverter
This can read a CSV format input stream and output a List.

The Spring MVC framework includes a mechanism for converting the raw
data contained in a HTTP request into equivalent Java or Kotlin class
instances. This mechanism can be extended to handle request data formats
that are not part of the default framework by providing an
implementation of the HttpMessageConverter interface. This commit
extends the framework to provide a MessageConverter implementation for
the Comma Separated Values format.

This implementation uses the Jackson JSON (or Jackson CSV) library to
perform the conversion. Although the framework provides an interface, it
also provides several partial implementations and extensions, in
particular the GenericHttpMessageConverter interface and the
AbstractHttpMessageConverter and AbstractGenericHttpMessageConverter
abstract classes. The last of these implements
GenericHttpMessageConverter and extends AbstractHttpMessageConverter and
is used as the basis for this commit's CsvHttpMessageConverter.

The use of GenericHttpMessageConverter arises from the record oriented
structure of CSV files. A List<something> (or Collection, Set...), where
something is a data class representation of the data in a row, is an
obvious representation of such a structure in Java or Kotlin. The
GenericHttpMessageConverter interface represents the ability to convert
some input into a sequence of items (somethings) stored in a 'generic'
datastructure and the AbstractGenericHttpMessageConverter class contains
the reusable framework code supporting this mechanism.