Add a PUT end-point for uploading a CSV file containing course data.
Now that we have an HttpMessageConverter that can convert CSV input to a
List representation in Kotlin we create the OpenAPI specification for
the PUT / courses end-point that accept a text/csv body which follows
the structure of the rows in the Courses spreadsheet:

name
description
audience
acronym
comments
As a result the code generation plugin adds this method to the
CoursesApiDelegate

fun coursesPut(coursesPutRequestInner: List<CoursesPutRequestInner>):
ResponseEntity<Unit>
and this commit adds an implementation of that method. Behind the scenes
Spring responds to the 'csv/text' Content-Type header by choosing the
CsvHttpMessageConverter to translate the uploaded CSV file into the
List<CoursesPutRequestInner> and the tests confirm that this behaves as
expected. A new CourseControllerTests method confirms this behaviour.