Refactor CoursesControllerTest
CourseControllerTest used to use an injected CourseSerivce which in turn
used InMemoryCourseRepository as a source of data. The controller now
calls a service method that takes a parameter, and the method
implementation is not trivial. I felt that it was better at this stage
to switch to using a mock implementation of CourseService instead with
the data for the mock responses being programmed from InMemoryRepository
as before.

Co-authored-by: Peter Bell <peter.bell@digital.justice.gov.uk>