Add a PUT /courses/prerequisites end-point to the Open API specification
This accepts prerequisite data as a bulk upload in CSV format.
All existing prerequisite data is to be deleted before the data in this upload is applied to the current set of courses.

The Gradle Open API Generator produces a skeleton implementation for the new end-point where the PrerequisiteRecord data class represents a row of data from the CSV file.