Split CoursesControllerTest into web-tier tests and a full stack integration tests
CoursesControllerTest began life as a test of CourseController class.
Most of the functionality of CourseController is provided through configuration,
convention and support from the Spring MVC framework so tests mainly confirm that
controller configuration is correct and interaction with the Spring Framework is
as expected.

As these are tests of CoursesController in the context of the MVC framework the tests
have no need of application services, repositories, databases etc and they do not
need to be configured.

Neverthelss these tests configured and tested the whole MVC - Service -
Repository - database stack. Mostly because this was easy to do and the
pre-populated database provided a useful source of test data. In fact,
initially there was no database, just an in-memory datastore built from Maps,
objects and Sets.

Then the InMemoryCourseRepositry was replaced by a database and a JPA implementation of
CourseRepository. This made the tests slower, but since the CourseService just delegated to
CourseRepository and the Repository layer was tested independently this still worked well.

Adding PUT methods to the API pushed this too far: The implementation of these methods changed
the state of the previously immutable database so that the tests were no longer
independent of ,each other.  Moreover, CourseService had acquired some logic of its own and
no-longer delegated all requests directly to CourseRepository.

The quick fix was to replace the CourseService injected into the controller with a mock
implementation.  The mock took data from InMemoryCourseRepository to feed to the Controller
and this approach was effective.

However, now the CourseControllerTests start a database, populate it, configure Hibernate
and wire everything together. This unnecessary, slow and wasteful.

Time to split CourseControllerTests into a set of full-stack integration tests and a
separate set of MVC tests.

CourseIntegrationTests now tests the full stack. It is not an exhaustive, since its purpose
 is to show that the components of the application work together.  Typically there is one
'happy flow' per api end-point.  Tests that alter the database state are marked as
@DirtiesContext which forces the test framework to re-configure the application before
running the next test.

CourseControllerTests now just configures and tests the MVC layer functionality
provided by CourseController.  These tests are faster and cover a wider range of scenarios.