Add a PUT /courses/offerings end-point to the Open API specification
This accepts offering data as a bulk upload in CSV format.
All existing offering data is to be deleted before the data in this upload is applied to the current set of courses.
CourseControllerTests are grouped by end-point using inner classes and @Nested support from JUnit 5.

The Gradle Open API Generator produces a skeleton implementation for the new end-point where the OfferingRecord data class represents a row of data from the CSV file.