Revert junit5spring to 4.6.0
We're currently seeing builds failing when tagging a Pact version on
`main` due to the deployed version not being published.

I'm not 100% sure if this is the culprit, but this commit was the first
one in which the `main` build started to fail, so I wanted to check this
first.