Add `secondaryContactEmail` to `CourseOffering`
It's come to light that there are multiple email addresses for some
CourseOfferings.

While we could have represented this by making the `contactEmail`
support a list of strings, it feels more appropriate to have this
secondary field as the vast majority of offerings only have a single
email, and having some kind of hierarchy of email addresses feels more
useful from a data perspective.