Refactor repository code for simplicity.
This PR contains a refactoring.  The application's functionality does not change.

The application initially implemented a read-only API. The domain model from which the API responses were constructed relied on a domain model containing instances of `CourseEntity`, `Offering` and `Prerequisite` classes. These entities were retrieved from a `CourseRepository`, an interface initially implemented wholly in memory as `InMemoryCourseRepository`.

Further development introduced a database as a long term store for these entities and a `JpaCourseRepository` implementation of `CourseRepository` as the replacement for `InMemoryCourseRepository`. `JpaCourseRepository` uses the Spring Data JPA support for Hibernate to implement the behaviour required of `CourseRepository`.

Subsequent iterations introduced a means of updating the data stored by the API. The necessary extensions to the `CourseRepository` interface were added to the domain as a separate interface, `MutableCourseRepository` which extended `CourseRepository`, and `JPACourseRepository` provided implementations of these new methods.  `InMemoryCourseRepository` was left unchanged but moved to the 'test' code tree as it provided a convenient source of data for certain tests.

This refactoring:
* Retains `InMemoryCourseRepository` for testing purposes.  It does not implement the `CourseRepository` interface, but its functionality is unchanged. `InMemoryCourseRepistoryTest` is untouched and all the tests still pass.
* Merges the `MutableCourseRepository` into the `CourseRepository` interface and updates `JpaCourseRepository` and its dependencies 'CourseService` etc accordingly.
* Moves `InMemoryCourseRepository` and its test class `InMemoryCourseRepositoryTest` from the `inmemoryrepo` package to the `restapi` package:  `InMemoryCourseRepository` is now only used by the `CourseControllerTest` class in the `restapi` package so it makes sense for the classes to sit along side each other.