Rewrite CourseService to update the courses instead of replacing them.
The intention is to preserve entity identities across bulk uploads by  mutating entities in place instead of replacing them. This includes withdrawing and then restoring withdrawn courses.

Operations are still mainly performed in memory.  This should be adequate until the number of withdrawn courses, offerings and prerequisites cannot be held in memory.

Replacing in-memory maps with database searches should be quite straightforward and would be the next step in handling large numbers of courses (possibly mostly withdrawn).