Refactor: Change the JPA relationship between `CourseEntity` and `Offering` from an `@ElementCollection` to a bi-directional `@OneToMany`.
This slightly complicates the code in `CourseEntity`, but enables an `OfferingRepository` for the Offering Entity.
This in turn simplifies the query for retrieving an Offering by Id, so overall I believe this is an improvement.
There is a case for making the same refactoring for the prerequisites collection, but there's no strong reason to make this change right now.