Updated latest src/main (V24__) and src/test (R__) flyway migrations for consistency.
Previous flyway migrations, which had been combined or deleted in places for clarity (namely V22-24),
now use different checksums. This may cause validation errors when running or debugging the application
using bootRunLocal. It is not possible to run flway repair directly in gradle due to the way we set up
out configuration, and for this ticket it's out-of-scope to include it. For reference, temporarily
registering a Bean which runs flyway.migrate() directly will allow the flyway_schema_history table to
update with the new checksums. This is not included in this commit or associated PR because it would
still be out of scope for this project, is generally only required once, seems not to affect the preprod
or production environments, and is somewhat memory-intensive to run.