Add Prison API Client and token generation classes
This adds a Prison API client which uses HMPPSBaseClient to make
requests to an external service. We're using it to fetch Sentence
Details and Sentence Dates, which will be exposed in the case list
endpoint.

HMPPS external services are secured through HMPPS Auth, which requires
clients to send an Oauth2 token. We're using Spring Security
Configuration to automagically generate this token and pass it to the
request. The credentials are fetched from the different deployment files
e.g. `values-dev.yaml` based on the deployed environment to ensure we're
fetching data from the correct environment.