Implement courseName filtering, and update supporting test framework.
Specifically, filtering is not case-sensitive and utilises the LIKE keyword in the database. Therefore, if, e.g., four referrals to "Red Course", "Yellow Course", "Green Course", and "Blue Course" exist for an organisation, and the only filter used was courseName with the value "course", all four referrals will have summaries built for them and returned as a list.