Update/tidy Pact tests
This updates the Pact tests to match the latest UI tests and removes
data that aren't being used anywhere. It ensures we're not
creating/updating/deleting entities we then rely on in later tests, as
state is persisted between tests, which has made this an incredibly
difficult thing to debug. We need to refactor this and reset state
between each test, otherwise these tests will continue to be very
brittle whenever we add/update them.

Co-authored-by: George <george@dxw.com>