Add missing pact states
The `Referral(s) exist for the logged in user` test is currently
failing due to our re-use of data between tests, which we urgently need
to fix:

In `R__test_data.sql`, we create one Referral for
`TEST_REFERRER_USER_1`. This is updated in tests to have a non-null
`submittedOn` value. In another test, we create a new referral, which
then has a null `submittedOn` value.

Pact [can't
handle](https://github.com/pact-foundation/pact-js/issues/393) these two
referrals having different types of their `submittedOn` values
(specifically one being `null` and one a `string`), so the expectation
fails.

Once we refactor our Pact tests to set up and tear down data in between
tests, we'll be able to write this test, and as we can't override the
username here due to the `@TestTemplate` config being shared across
tests with the same referrer username. For this reason, we're going to
disable that individual assertion on the UI side for now, but I've left
this marked as `TODO` to remind us to add this assertion and re-enable
it on the UI side.