Stub Prisoner Search with Wiremock in Pact tests
Because we weren't stubbing the Prisoner Search here, we were getting
back `null` values for all Prisoners returned. Pact was failing due to
the consumer tests expecting these to be strings.

This also fixes some test constants, using Prison Numbers that match
those found in the `test_data.sql` file. These are a bit messy at the
moment and need tidying up, but that can come in a separate PR.