refactor(ReferralController) Add documentation to the getNextStatusTransitions endpoint
**WHAT**

This is a pure-refactor commit, i.e. it introduces no new functionality.

Made the following changes to the getNextStatusTransitions endpoint (i.e. `GET /referrals/{id}/status-transitions`):

- Added the `@Operation` annotation
- Replaced the `@GetMapping` annotation with the `@RequestMapping`
- Return a 404 if the Referral isn't found (don't coerce non-null)
- Refactored the logic to stop re-finding of the same status (i.e. the DESELECTED code)

**WHY**

As part of getting more familiar with Kotlin and the AcP codebase(s), I have been getting familiar with the patterns used.

While looking around, it appeared that this endpoint was broadly inconsistent with the others.