fix(SexualOffenceDetails): Update documentation and types
**WHAT**

- Update swagger documentation of the SexualOffenceDetails
- Update the SexualOffenceDetails::categoryCode to be an enum

**WHY**

The SexualOffenceDetails documentation was a little copy-pasted, and now
has realistic looking data in the annotations which will improve
developer experience in code, and also of those using our OpenAPI spec.

The SexualOffenceDetails::categoryCode field is set entirely from SQL
migrations, and as thus has a fixed and known subset of values, for
which an enum _exists_ but was not used in the data definition.  This
change helps make the data model a little more representative of
reality.